deleting the file

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require("../models/product");
 const mongoose = require("mongoose");
 const { validationResult } = require("express-validator");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -157,7 +158,10 @@ exports.postEditProduct = (req, res, next) => {
 
       product.title = updatedTitle;
       product.description = updatedDesc;
-      if (image) product.imageUrl = image.path;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.price = updatedPrice;
 
       return product.save().then((result) => {
@@ -175,7 +179,15 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
 
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product Not Found"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("Product Deleted");
       return res.redirect("/admin/products");
